Admin fixes pt1
