+++ dapps/admin/README.md
@@ -1,32 +1,30 @@
-![Origin Protocol](data/origin-header.png)
+# ![Origin Protocol](data/origin-header.png)
 
 A UI leveraging `@origin/graphql`. View and manage listings and offers.
 
-Demo at https://www.originadm.in
+Demo at [https://ogn.dev/admin](https://ogn.dev/admin)
 
 ## Usage
 
-    npx lerna bootstrap
-    npm start
-
-`npm start` also starts a local blockchain and IPFS server.
+    # Run from the monorepo root
+    yarn
+    cd dapps/admin
+    yarn start
 
 Watch the [video tutorial](https://drive.google.com/a/originprotocol.com/file/d/1JHXtYzl7qFyTNB62lNlOwZ_L6QCQnydB/view?usp=sharing) to see how to use this UI.
 
 ## Troubleshooting
 
-### How do I test moderation localy?
+### Testing moderation localy
 
 Go to `packages/graphql/src/configs/localhost.js` and uncomment the discovery server.
 
-### How do I reset my local blockchain?
+### Resetting local data
 
 From the top level of your local repo:
 
-```
-cd packages/services
-num run clean
-```
+    cd packages/services
+    yarn run clean
 
 ### The UI is stuck on "loading..."
