+++ dapps/admin/package.json
@@ -34,7 +34,6 @@
     "@origin/services": "^0.1.0",
     "apollo-link": "^1.2.12",
     "babel-plugin-module-resolver": "^3.2.0",
-    "body-parser": "^1.19.0",
     "bs58": "^4.0.1",
     "date-fns": "^2.0.0",
     "express": "^4.17.1",
