+++ dapps/admin/src/pages/settings/_RemoveWalletBtn.js
